#4 - Add support to change font and font size for each text passage and header and footer - fix not working font-size - use double instead of float to avoid cast - extend a few code documentations - use "content" as argument name in a few places to better reflect argument kind - fix missing code highlight in markdown documentation

+++ NetOdt/Helper/FontHelper.cs
@@ -12,7 +12,7 @@ namespace NetOdt.Helper
         /// </summary>
         /// <param name="fontSize">The font size representation for the real font size</param>
         /// <returns>A real font size</returns>
-        internal static float GetFontSize(FontSize fontSize)
-            => fontSize == FontSize.Size10Point5 ? 10.5f : (float)fontSize;
+        internal static double GetFontSize(FontSize fontSize)
+            => fontSize == FontSize.Size10Point5 ? 10.5 : (double)fontSize;
     }
 }
